corrected syscall types

+++ libk/src/syscall/syscalls.h
@@ -5,5 +5,5 @@
 
 int syscall_open_file(const char* path, int mode);
 bool syscall_close_file(int fd);
-int syscall_read_file(int fd, char* buff, size_t count);
-int syscall_write_file(int fd, char* buff, size_t count);
+size_t syscall_read_file(int fd, char* buff, size_t count);
+size_t syscall_write_file(int fd, char* buff, size_t count);
